+++ memsource/__init__.py
@@ -1,3 +1,3 @@
 __author__ = 'Gengo'
-__version__ = '0.5.1'
+__version__ = '0.5.2'
 __license__ = 'MIT'